Add "remote-branch" config parameter
Fixes cases where the default branch of a remote repository with the
schemes is not "master" but "main" for example.
We still default to "master".

Adding a config parameter was the best option in my opinion. Ideally the
default branch would be given in the templates' list but we're not the
owners of that list. The easiest is to add a config parameter to
specify the branch to use