Add "remote-branch" config parameter & handle scheme links failing with 404 (#63)
* Add "remote-branch" config parameter

Fixes cases where the default branch of a remote repository with the
schemes is not "master" but "main" for example.
We still default to "master".

Adding a config parameter was the best option in my opinion. Ideally the
default branch would be given in the templates' list but we're not the
owners of that list. The easiest is to add a config parameter to
specify the branch to use

* Syntax / Whitespace changes

* git:Ignore JetBrains .idea/ folder

* helpers:Stop panicking when getting 404 from color scheme

The current scheme is not available:
    https://github.com/aramisgithub/base16-ia-scheme

Pull request to remove the links to that non existing user:
    https://github.com/chriskempson/base16-schemes-source/pull/7